improved file processing and exiting

+++ flac_diag.sh
@@ -15,34 +15,28 @@ check_requisites () {
 		if [[ -z $(command -v $package) ]]; then
 			echo 'The following program is not installed or cannot be found in this users $PATH: '$package
 			echo 'Fix it and try again.'
-			end_bad
+			end "Missing packages. Cannot continue." 1
 		fi
 	done
 }
 
-end_bad () {
+# message then status
+end () {
 	cleanup
-	echo '###############################################'
-	echo 'REACHED THE END OF THE SCRIPT with AN ERROR!'
-	echo '###############################################'
-	exit 1
+	echo '################################################'
+	echo 'END OF THE SCRIPT'
+	echo 'MESSAGE: '$1 
+	echo '################################################'
+	exit $2
 }
 
-end_good () {
-	cleanup
-	echo '###############################################'
-	echo 'REACHED THE END OF THE SCRIPT without ERRORS.'
-	echo '###############################################'
-	exit 0
-}
-
-end_int () {
+interrupt () {
 	echo '!! ATTENTION !!'
-	echo 'Received a signal to stop the program right now.'
+	echo 'Received a signal to stop the script right now.'
 	cleanup
-	echo '###############################################'
-	echo 'THE SCRIPT REACHED A CLEAN STOP.'
-	echo '###############################################'
+	echo '############################################'
+	echo '##### THE SCRIPT REACHED A CLEAN STOP. #####'
+	echo '############################################'
 	exit 1
 }
 
@@ -51,8 +45,7 @@ prepare_test_flac () {
 	DIR="$1"
 	if [[ -z $DIR || ! -d $DIR ]]; then
 		echo 'No directory was provided or the argument is not a directory.'
-		echo 'Please provide the full path to a directory when running this script.'
-		end_bad
+		end "Please provide the full path to a directory when running this script." 1
 	fi
 	# prepare log folders and files
 	LOG_FOLDER='./log/'
@@ -107,7 +100,7 @@ run_test_flac () {
 	echo 'Starting test...'
 	# find flac files recursively and do something with them (one at a time)
 	find "$DIR" -iname '*.flac' -print0 | while read -d $'\0' file; do
-		# skip if this file has been analyzed before
+		# skip files that have been analyzed before
 		if [[ $(cat $GOOD_LOG | grep -F "$file") ]]; then
 			echo '---------------'
 			echo 'Skipping' $file
@@ -120,30 +113,53 @@ run_test_flac () {
 			echo 'This file has already been processed before and it was BAD then.'
 			echo '---------------'
 			continue
+		# check file that is not accessible anymore
+		elif [[ ! -f $file ]]; then
+			echo '---------------'
+			echo 'Skipping' $file
+			echo 'This is awkward because the file does not seem to exist anymore.'
+			if [[ ! -d $DIR ]]; then
+				echo 'It looks like the folder' $DIR 'has been removed/unmounted.'
+				echo 'No point in continuing.'
+				echo '---------------'
+				end "Cannot read files anymore" 1
+			else
+				echo 'The folder' $DIR 'looks accessible. If there are mounts in this folder, you might have lost connection to them.'
+			fi
+			echo 'Will skip this file but will try the next one anyway.'
+			echo '---------------'
+			continue
 		fi
-		# process file else
+		# process file
 		echo '---------------'
 		echo 'Processing file:' $file
 		echo 'Date and time:' $(date)
 		flac -st "$file" > $CACHE 2>&1
 		FLAC=$(cat $CACHE)
 		if [[  $FLAC  ]]; then
-			echo 'Uh-oh! THE FLAC FILE HAS AN ERROR!'
-			FILE_FLAC_VERSION=$(metaflac --show-vendor-tag "$file" 2>&1 | grep -oE $REGEX_FLAC_VERSION | tr -dc '0-9')
-			if [[ -z $FILE_FLAC_VERSION ]]; then
-				echo 'I am unable to retrieve the flac version from the file.'
-			elif [[ $FLAC_VERSION < $FILE_FLAC_VERSION ]]; then
-				echo 'You are possibly using an OUTDATED FLAC VERSION.'
-				echo 'Try to update your flac and run this script again after cleaning' $BAD_LOG '.'
-			elif [[ $FLAC_VERSION > $FILE_FLAC_VERSION ]]; then
-				echo 'The audio file is LIKELY CORRUPTED.'
+			# catch errors due to file not being accessible anymore
+			if [[ -f $file ]]; then
+				echo 'Uh-oh! THE FLAC FILE HAS AN ERROR!'
+				FILE_FLAC_VERSION=$(metaflac --show-vendor-tag "$file" 2>&1 | grep -oE $REGEX_FLAC_VERSION | tr -dc '0-9')
+				if [[ -z $FILE_FLAC_VERSION ]]; then
+					echo 'I am unable to retrieve the flac version from the file.'
+				elif [[ $FLAC_VERSION < $FILE_FLAC_VERSION ]]; then
+					echo 'You are possibly using an OUTDATED FLAC VERSION.'
+					echo 'Try to update your flac and run this script again after cleaning' $BAD_LOG '.'
+				elif [[ $FLAC_VERSION > $FILE_FLAC_VERSION ]]; then
+					echo 'The audio file is LIKELY CORRUPTED.'
+				fi
+				echo 'The file will be added to' $BAD_LOG
+				echo $file >> $BAD_LOG
+				echo 'To investigate the error, check' $CACHE_ERRORS
+				FILENAME=$(echo $file | grep -oE $REGEX_FILENAME)
+				ERROR_FILE=$CACHE_ERRORS$FILENAME'.err'
+				cat $CACHE > "$ERROR_FILE"
+			else
+				echo 'This file is NO LONGER AVAILABLE.'
+				echo 'Will try the next one.'
+				continue
 			fi
-			echo 'The file will be added to' $BAD_LOG
-			echo $file >> $BAD_LOG
-			echo 'To investigate the error, check' $CACHE_ERRORS
-			FILENAME=$(echo $file | grep -oE $REGEX_FILENAME)
-			ERROR_FILE=$CACHE_ERRORS$FILENAME'.err'
-			cat $CACHE > "$ERROR_FILE"
 		else
 			echo 'Good news, everyone! THE FLAC FILE IS OKAY!'
 			echo 'The file will be added to' $GOOD_LOG
@@ -171,6 +187,8 @@ start () {
 start
 check_requisites
 prepare_test_flac "$1"
-trap 'end_int' SIGINT SIGTERM SIGKILL
+trap 'interrupt' SIGINT SIGHUP SIGTERM SIGKILL
 run_test_flac
-end_good
+if [[ $? = 0 ]]; then
+	end "The script finished without errors." 0
+fi
